Added QSettings. Moved dialog classes for settings in its own module.
